CRUD əməliyyatları hazırdır

+++ spring/course/course/src/main/java/az/developia/course/service/ParentService.java
@@ -1,9 +1,12 @@
 package az.developia.course.service;
 
-import java.util.List; 
+import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import az.developia.course.exception.MyBadRequestException;
 import az.developia.course.model.Parent;
 import az.developia.course.repository.ParentRepository;
 
@@ -30,8 +33,24 @@ public class ParentService {
 	
 	public void deleteById(Integer id) {
 		
-		parentRepository.deleteById(id);
+		Optional<Parent> findedStudent = parentRepository.findById(id);
+		if (findedStudent.isPresent()) {
+			
+			parentRepository.deleteById(id);
+		}else {
+			throw new MyBadRequestException("id tapılmadı");
+		}
+		
+		
 	}
-	
+
+
+
+	public Optional<Parent> findById(Integer id) {
+		
+		return parentRepository.findById(id);
+	}
+
+
 
 }
