+++ spring/course/course/src/main/java/az/developia/course/service/StudentService.java
@@ -1,10 +1,12 @@
 package az.developia.course.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import az.developia.course.exception.MyBadRequestException;
 import az.developia.course.model.Student;
 import az.developia.course.repository.StudentRepository;
 
@@ -31,7 +33,22 @@ public class StudentService {
 	
 	public void deleteById(Integer id) {
 		
-		studentRepository.deleteById(id);
+		Optional<Student> findedStudent = studentRepository.findById(id);
+		if (findedStudent.isPresent()) {
+			
+			studentRepository.deleteById(id);
+		}else {
+			throw new MyBadRequestException("id tapılmadı");
+		}
+		
+		
+	}
+
+
+
+	public Optional<Student> findById(Integer id) {
+		
+		return studentRepository.findById(id);
 	}
