+++ spring/course/course/src/main/java/az/developia/course/service/TeacherService.java
@@ -1,9 +1,12 @@
 package az.developia.course.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import az.developia.course.exception.MyBadRequestException;
 import az.developia.course.model.Teacher;
 import az.developia.course.repository.TeacherRepository;
 
@@ -30,7 +33,21 @@ public class TeacherService {
 	
 	public void deleteById(Integer id) {
 		
-		teacherRepository.deleteById(id);
+		Optional<Teacher> findedStudent = teacherRepository.findById(id);
+		if (findedStudent.isPresent()) {
+			
+			teacherRepository.deleteById(id);
+		}else {
+			throw new MyBadRequestException("id tapılmadı");
+		}
+		
+	
+	}
+
+	
+	public Optional<Teacher> findById(Integer id) {
+		
+		return teacherRepository.findById(id);
 	}
